[MS-1229] move review fetch logic into separate hook
